+++ tests/test_feed_storage_module/test_baleage.py
@@ -2,7 +2,7 @@ import pytest
 from pytest_mock import MockerFixture
 
 from RUFAS.routines.feed_storage.baleage import Baleage, INITIAL_LOSS_PERIOD, DEFAULT_FINAL_MOISTURE_PERCENTAGE
-from RUFAS.data_structures.crop_soil_feed_storage_connection import CropCategory, CropType, HarvestedCrop
+from RUFAS.data_structures.crop_soil_to_feed_storage_connection import CropCategory, CropType, HarvestedCrop
 from RUFAS.time import Time
 
 from .sample_crop_data import sample_crop_data
