+++ tests/test_feed_storage_module/test_grain.py
@@ -1,6 +1,6 @@
 import pytest
 
-from RUFAS.data_structures.crop_soil_feed_storage_connection import CropCategory
+from RUFAS.data_structures.crop_soil_to_feed_storage_connection import CropCategory
 from RUFAS.routines.feed_storage.grain import Grain
