+++ tests/test_feed_storage_module/test_storage.py
@@ -6,7 +6,7 @@ from pytest_mock import MockerFixture
 
 from RUFAS.current_day_conditions import CurrentDayConditions
 from RUFAS.general_constants import GeneralConstants
-from RUFAS.data_structures.crop_soil_feed_storage_connection import CropCategory, CropType, HarvestedCrop
+from RUFAS.data_structures.crop_soil_to_feed_storage_connection import CropCategory, CropType, HarvestedCrop
 from RUFAS.routines.feed_storage.storage import Storage
 from RUFAS.time import Time
 from RUFAS.units import MeasurementUnits
@@ -225,7 +225,7 @@ def test_get_total_nutritive_amount(
     """Test _get_total_nutritive_amount in Storage class."""
     storage.stored = [harvested_crop] * len(percentages)
     mock_dry_matter_mass = mocker.patch(
-        "RUFAS.data_structures.crop_soil_feed_storage_connection.HarvestedCrop.dry_matter_mass",
+        "RUFAS.data_structures.crop_soil_to_feed_storage_connection.HarvestedCrop.dry_matter_mass",
         new_callable=mocker.PropertyMock,
         return_value=dry_mass,
     )
@@ -265,7 +265,7 @@ def test_calculate_dry_matter_loss_to_gas(
 ) -> None:
     """Tests calculate_dry_matter_loss_to_gas in Storage."""
     mocker.patch(
-        "RUFAS.data_structures.crop_soil_feed_storage_connection.HarvestedCrop.dry_matter_mass",
+        "RUFAS.data_structures.crop_soil_to_feed_storage_connection.HarvestedCrop.dry_matter_mass",
         new_callable=mocker.PropertyMock,
         return_value=dry_matter,
     )
